+++ src/main/java/io/github/astrapi69/collection/list/SortedUniqueList.java
@@ -26,44 +26,65 @@ package io.github.astrapi69.collection.list;
 
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Collections;
+import java.util.Comparator;
 
 /**
- * The class {@link UniqueArrayList} extends the {@link ArrayList} and overwrites all add-methods
- * that checks if the element already exists in the list
+ * The class {@link SortedUniqueList} extends the {@link ArrayList} and ensures that all elements
+ * are unique and sorted according to the given comparator
  *
  * @param <E>
  *            the generic type of values
- * @see java.util.ArrayList
  */
-public class UniqueArrayList<E> extends ArrayList<E>
+public class SortedUniqueList<E> extends ArrayList<E>
 {
 
 	/**
-	 * Constructs an empty {@link UniqueArrayList} instance with the specified initial capacity
+	 * The comparator for sorting this list object
+	 */
+	private Comparator<? super E> comparator;
+
+	/**
+	 * Constructs an empty {@link SortedUniqueList} instance with the specified initial capacity
 	 *
 	 * @param initialCapacity
 	 *            the initial capacity
 	 */
-	public UniqueArrayList(int initialCapacity)
+	public SortedUniqueList(int initialCapacity)
 	{
 		super(initialCapacity);
 	}
 
 	/**
-	 * Constructs an empty {@link UniqueArrayList} instance with an initial capacity of ten
+	 * Constructs an empty {@link SortedUniqueList} instance with an initial capacity of ten
 	 */
-	public UniqueArrayList()
+	public SortedUniqueList()
 	{
 	}
 
 	/**
-	 * Constructs a {@link UniqueArrayList} instance containing the elements of the given collection
+	 * Constructs a {@link SortedUniqueList} instance and sets the given comparator
+	 *
+	 * @param comparator
+	 *            the comparator
+	 */
+	public SortedUniqueList(Comparator<? super E> comparator)
+	{
+		this.comparator = comparator;
+	}
+
+	/**
+	 * Constructs a {@link SortedUniqueList} instance containing the elements of the given
+	 * collection and sets the given comparator
 	 *
 	 * @param collection
 	 *            the collection
+	 * @param comparator
+	 *            the comparator
 	 */
-	public UniqueArrayList(Collection<? extends E> collection)
+	public SortedUniqueList(Collection<? extends E> collection, Comparator<? super E> comparator)
 	{
+		this(comparator);
 		addAll(collection);
 	}
 
@@ -75,7 +96,11 @@ public class UniqueArrayList<E> extends ArrayList<E>
 	{
 		if (!contains(element))
 		{
-			return super.add(element);
+			int insertIndex = comparator != null
+				? ListExtensions.getIndexToInsert(this, element, comparator)
+				: size();
+			super.add(insertIndex, element);
+			return true;
 		}
 		return false;
 	}
@@ -86,17 +111,7 @@ public class UniqueArrayList<E> extends ArrayList<E>
 	@Override
 	public void add(int index, E element)
 	{
-		if (contains(element))
-		{
-			if (remove(element))
-			{
-				super.add(index, element);
-			}
-		}
-		else
-		{
-			super.add(index, element);
-		}
+		add(element);
 	}
 
 	/**
@@ -105,15 +120,12 @@ public class UniqueArrayList<E> extends ArrayList<E>
 	@Override
 	public boolean addAll(Collection<? extends E> collection)
 	{
-		if (collection.size() == 0)
-		{
-			return false;
-		}
+		boolean result = false;
 		for (E element : collection)
 		{
-			this.add(element);
+			result |= add(element);
 		}
-		return true;
+		return result;
 	}
 
 	/**
@@ -122,23 +134,28 @@ public class UniqueArrayList<E> extends ArrayList<E>
 	@Override
 	public boolean addAll(int index, Collection<? extends E> collection)
 	{
-		if (collection.size() == 0)
-		{
-			return false;
-		}
-		int currentIndex = index;
-		for (E element : collection)
-		{
-			if (currentIndex >= size())
-			{
-				this.add(element);
-			}
-			else
-			{
-				this.add(currentIndex, element);
-			}
-			currentIndex++;
-		}
-		return true;
+		return addAll(collection);
+	}
+
+	/**
+	 * Gets the comparator of this list
+	 *
+	 * @return the comparator of this list
+	 */
+	public Comparator<? super E> getComparator()
+	{
+		return comparator;
+	}
+
+	/**
+	 * Sets the comparator of this list
+	 *
+	 * @param comparator
+	 *            the new comparator of this list
+	 */
+	public void setComparator(Comparator<? super E> comparator)
+	{
+		this.comparator = comparator;
+		Collections.sort(this, comparator);
 	}
 }