+++ src/main/java/io/github/astrapi69/collection/list/UniqueList.java
@@ -26,66 +26,44 @@ package io.github.astrapi69.collection.list;
 
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
-import java.util.Comparator;
 
 /**
- * The class {@link SortedArrayList} extends the {@link ArrayList} and overwrites all add-methods
- * that sorts the elements with the given comparator.
+ * The class {@link UniqueList} extends the {@link ArrayList} and overwrites all add-methods that
+ * checks if the element already exists in the list
  *
  * @param <E>
  *            the generic type of values
- * @see ArrayList
+ * @see java.util.ArrayList
  */
-public class SortedArrayList<E> extends ArrayList<E>
+public class UniqueList<E> extends ArrayList<E>
 {
 
 	/**
-	 * the comparator for sort this list object
-	 */
-	private Comparator<? super E> comparator;
-
-	/**
-	 * Constructs an empty {@link SortedArrayList} instance with the specified initial capacity
+	 * Constructs an empty {@link UniqueList} instance with the specified initial capacity
 	 *
 	 * @param initialCapacity
 	 *            the initial capacity
 	 */
-	public SortedArrayList(int initialCapacity)
+	public UniqueList(int initialCapacity)
 	{
 		super(initialCapacity);
 	}
 
 	/**
-	 * Constructs an empty {@link SortedArrayList} instance with an initial capacity of ten
+	 * Constructs an empty {@link UniqueList} instance with an initial capacity of ten
 	 */
-	public SortedArrayList()
+	public UniqueList()
 	{
 	}
 
 	/**
-	 * Constructs a {@link SortedArrayList} instance and set the given comparator
-	 *
-	 * @param comparator
-	 *            the comparator
-	 */
-	public SortedArrayList(Comparator<? super E> comparator)
-	{
-		this.comparator = comparator;
-	}
-
-	/**
-	 * Constructs a {@link SortedArrayList} instance containing the elements of the given collection
-	 * and set the given comparator
+	 * Constructs a {@link UniqueList} instance containing the elements of the given collection
 	 *
 	 * @param collection
 	 *            the collection
-	 * @param comparator
-	 *            the comparator
 	 */
-	public SortedArrayList(Collection<? extends E> collection, Comparator<? super E> comparator)
+	public UniqueList(Collection<? extends E> collection)
 	{
-		this(comparator);
 		addAll(collection);
 	}
 
@@ -95,15 +73,11 @@ public class SortedArrayList<E> extends ArrayList<E>
 	@Override
 	public boolean add(E element)
 	{
-		int currentSize = this.size();
-		if (comparator == null)
+		if (!contains(element))
 		{
 			return super.add(element);
 		}
-		ListExtensions.getIndexToInsert(this, element, comparator);
-		int insertIndex = ListExtensions.getIndexToInsert(this, element, comparator);
-		super.add(insertIndex, element);
-		return this.size() != currentSize;
+		return false;
 	}
 
 	/**
@@ -112,7 +86,17 @@ public class SortedArrayList<E> extends ArrayList<E>
 	@Override
 	public void add(int index, E element)
 	{
-		add(element);
+		if (contains(element))
+		{
+			if (remove(element))
+			{
+				super.add(index, element);
+			}
+		}
+		else
+		{
+			super.add(index, element);
+		}
 	}
 
 	/**
@@ -121,12 +105,15 @@ public class SortedArrayList<E> extends ArrayList<E>
 	@Override
 	public boolean addAll(Collection<? extends E> collection)
 	{
-		boolean result = false;
+		if (collection.size() == 0)
+		{
+			return false;
+		}
 		for (E element : collection)
 		{
-			result |= add(element);
+			this.add(element);
 		}
-		return result;
+		return true;
 	}
 
 	/**
@@ -135,28 +122,23 @@ public class SortedArrayList<E> extends ArrayList<E>
 	@Override
 	public boolean addAll(int index, Collection<? extends E> collection)
 	{
-		return addAll(collection);
-	}
-
-	/**
-	 * Gets the comparator of this list
-	 * 
-	 * @return the comparator of this list
-	 */
-	public Comparator<? super E> getComparator()
-	{
-		return comparator;
-	}
-
-	/**
-	 * Sets the comparator of this list
-	 * 
-	 * @param comparator
-	 *            the new comparator of this list
-	 */
-	public void setComparator(Comparator<? super E> comparator)
-	{
-		this.comparator = comparator;
-		Collections.sort(this, comparator);
+		if (collection.size() == 0)
+		{
+			return false;
+		}
+		int currentIndex = index;
+		for (E element : collection)
+		{
+			if (currentIndex >= size())
+			{
+				this.add(element);
+			}
+			else
+			{
+				this.add(currentIndex, element);
+			}
+			currentIndex++;
+		}
+		return true;
 	}
 }