+++ src/main/java/io/github/astrapi69/collection/list/package-info.java
@@ -1,4 +1,4 @@
 /**
- * This package provides utility class for list objects
+ * This package provides utility classes for list and vector objects
  */
 package io.github.astrapi69.collection.list;