+++ src/main/java/io/github/astrapi69/collection/map/package-info.java
@@ -1,4 +1,4 @@
 /**
- * This package provides utility class for map objects
+ * This package provides utility classes for map objects and map implementations
  */
 package io.github.astrapi69.collection.map;